pyproject.toml
[tool.poetry]
name = "python-ldraw"
version = "0.8.2"
description = "Python wrapper for LDraw library."
authors = ["David Boddie <david@boddie.org.uk>", "Matthieu Berthomé <rienafairefr@gmail.com>"]
packages = [ { include = "ldraw" } ]

[tool.poetry.scripts]
ldr2inv = "ldraw.tools.ldr2inv:main"
ldr2png = "ldraw.tools.ldr2png:main"
ldr2pov = "ldraw.tools.ldr2pov:main"
ldr2svg = "ldraw.tools.ldr2svg:main"
ldraw = "ldraw.cli:main"


[tool.poetry.dependencies]
python = "^3.6.1"
appdirs = "^1.4.4"
numpy = "^1.19.0"
pymklist = "==1.1.0"
pystache = "^0.5.4"
attrdict = "^2.0.1"
progress = "^1.5"
pyyaml = "^5.3.1"
pillow = "^7.2.0"
mock = "^4.0.3"
requests = "^2.25.1"
requests-cache = "^0.5.2"
inquirer = "^2.7.0"
inflect = "^5.0.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.6.0"
black = "^19.10b0"
tox = "^3.16.1"
pytest = "^5.4.3"
future = "^0.18.2"
pytest-cov = "^2.10.1"
setuptools-scm = "^5.0.1"
setuptools = "^51.1.1"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (run with '--integration')",
]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
